svn path=/trunk/yarp2/; revision=8682

+++ src/carriers/xmlrpc_carrier/xmlrpc/XmlRpcServerConnection.cpp
@@ -371,7 +371,7 @@ XmlRpcServerConnection::generateHeader(std::string const& body)
     "Content-length: ";
 
   char buffLen[40];
-  sprintf(buffLen,"%d\r\n\r\n", body.size());
+  sprintf(buffLen,"%d\r\n\r\n", (int)body.size());
 
   return header + buffLen;
 }
